% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_essential.R
\name{et_essential}
\alias{et_essential}
\title{et_essential}
\usage{
et_essential(tool, input_df, cutoff)
}
\arguments{
\item{tool}{Tool used to generate the results; one of Gumbel or Tradis.}

\item{input_df}{Data frame containing essentiality call for all replicates
for a given condition.}

\item{cutoff}{Threshold for determining essentiality based on the number of
replicates.}
}
\value{
Data frame of essential genes for the condition.
}
\description{
et_essential
}
\references{
None.
}
\seealso{
\url{https://github.com/travis-m-blimkie/EssentialTnSeq}
}
