% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_renameColumns.R
\name{et_renameColumns}
\alias{et_renameColumns}
\title{et_renameColumns}
\usage{
et_renameColumns(input_list, condition)
}
\arguments{
\item{input_list}{Named list of data frames for a given condition, with each
data frame corresponding to a different replicate.}

\item{condition}{Condition name (string) for the given list.}
}
\value{
The same list of data frames, but with columns renamed to be specific
  to the condition.
}
\description{
Renames columns of all data frames in a list based on the name
  of the list element containing those data frames. Designed to use the
  output from \code{EssentialTnSeq::et_readFiles()}.
}
\examples{
\dontrun{
  et_renameColumns(
    input_list = treatment1_list,
    condition = "treatment1"
  )
}

}
\references{
None.
}
\seealso{
\url{https://github.com/hancockinformatics/EssentialTnSeq}
}
