% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_essential.R
\name{true_essential}
\alias{true_essential}
\title{true_essential}
\usage{
true_essential(dfs, tool, total_num)
}
\arguments{
\item{dfs}{List of data frames for all conditions/replicates}

\item{tool}{Tool that was used to genrate essentiality data; currently
supports either Tradis or Transit/Gumbel}

\item{total_num}{Total number of conditions/replicates}
}
\description{
Based on a set of data frames containing TnSeq results (from
  Tradis or Transit/Gumbel), returns genes which are condidered essential in
  all conditions/replicates.
}
\examples{

}
