% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_essential.R
\name{et_essential}
\alias{et_essential}
\title{et_essential}
\usage{
et_essential(tool, input_df, cutoff)
}
\arguments{
\item{tool}{String; one of "Gumbel" or "Tradis".}

\item{input_df}{Data frame containing essentiality call for all replicates
for a given condition.}

\item{cutoff}{Threshold for determining essentiality based on the number of
replicates. Recommend setting to one less than the number of replicates.}
}
\value{
Filtered data frame of essential genes for the condition.
}
\description{
Determines essentiality for a set of genes from Gumbel or
  Tradis. For Gumbel, a gene is considered essential if it was assigned a
  call of "E". For Tradis, essential genes are those with 0 read counts. The
  filtered data frame which is returned contains the additional columns used
  in defining essentiality.
}
\examples{
\dontrun{
  et_essential("Gumbel", treatment1_df, cutoff = 2)
}

}
\references{
None.
}
\seealso{
\url{https://github.com/travis-m-blimkie/EssentialTnSeq}
}
