% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_readFiles.R
\name{et_readFiles}
\alias{et_readFiles}
\title{et_readFiles}
\usage{
et_readFiles(tool, conditions, reps, data_folder)
}
\arguments{
\item{tool}{String; one of "Gumbel" or "Tradis".}

\item{conditions}{List of conditions. These should correspond to file names,
and should be spcific and non-overlapping.}

\item{reps}{Number of replicates, also corresponding to the number of files
for each condition}

\item{data_folder}{Directory containing files for all conditionas and
replicates.}
}
\value{
Nested and named list of data frames for all conditions and
  replicates.
}
\description{
Reads in multiples files, corresponding to different conditions
  and replicates from TnSeq analysis with Gumbel of Tradis. Creates a nested,
  named list of data frames for further analysis.
}
\references{
None.
}
\seealso{
\url{https://github.com/travis-m-blimkie/EssentialTnSeq}
}
